same fix as in bpll before

+++ python3/pracmln/mln/learning/cll.py
@@ -280,7 +280,7 @@ class CLL(AbstractLearner):
                 for i, val in enumerate(values):
                     v -= self.probs[part][i] * val
                 grad[fidx] += v
-        self.grad_opt_norm = float(sqrt(float(fsum([x * x for x in grad]))))
+        self.grad_opt_norm = sqrt(float(fsum([x * x for x in grad])))
         return numpy.array(grad)
